feat: Lighting Panel Setup Support

+++ setup_wizard/character_rig_setup/rig_script.py
@@ -6,6 +6,8 @@ import os
 from mathutils import Color, Vector
 from math import pi
 
+from setup_wizard.geometry_nodes_setup.lighting_panel_names import LightingPanelNames
+
 def rig_character(
         file_path, 
         disallow_arm_ik_stretch, 
@@ -853,7 +855,7 @@ def rig_character(
     bpy.data.objects["metarig"].hide_render = True
 
     # Moves specified param and it's children into the collection
-    def move_into_collection(object,collection):
+    def move_into_collection(object,collection,include_children=True):
         
         # Get object
         this_obj = bpy.context.scene.objects.get(object)
@@ -872,12 +874,13 @@ def rig_character(
             # Move it to our specified collection (This only does the parent obj)
             this_coll.objects.link(this_obj)
             
-            # Now we move the children of this object
-            for child_obj in this_obj.children:
-                # Same thing, unlink previous collections
-                for coll in child_obj.users_collection:
-                    coll.objects.unlink(child_obj)
-                this_coll.objects.link(child_obj)
+            if include_children:
+                # Now we move the children of this object
+                for child_obj in this_obj.children:
+                    # Same thing, unlink previous collections
+                    for coll in child_obj.users_collection:
+                        coll.objects.unlink(child_obj)
+                    this_coll.objects.link(child_obj)
                 
     # Move the rig into the char name's collection                        
     move_into_collection(char_name+"Rig",char_name)
@@ -893,6 +896,7 @@ def rig_character(
             coll.name = "WG"
 
     move_into_collection("metarig","wgt")
+
 
     # Unlink all inner objects from the old WGT collection. We want them inside the new one.
     for obj in bpy.data.objects:
@@ -934,22 +938,30 @@ def rig_character(
     head_up_obj.hide_viewport = True
     head_up_obj.hide_render = True
 
+    # IMPORTANT: This must be done before deleting the "Collection" collection in case Lighting Panel gets appended in there
+    # remove lighting colls - also move the RGB wheels into the rig obj
+    lighting_panel_rig_obj = bpy.data.objects.get(LightingPanelNames.Objects.LIGHTING_PANEL)
+    if lighting_panel_rig_obj:
+        to_del_coll = bpy.data.collections.get(LightingPanelNames.Collections.WIDGET_COLLECTION)
+        for obj in to_del_coll.objects:
+            move_into_collection(obj.name, "wgt")
+        to_del_coll = bpy.data.collections.get(LightingPanelNames.Collections.PICKER)
+        for obj in to_del_coll.objects:
+            move_into_collection(obj.name, "wgt")
+        to_del_coll = bpy.data.collections.get(LightingPanelNames.Collections.WHEEL)
+        for obj in to_del_coll.objects:
+            move_into_collection(obj.name, char_name)
+        # DO NOT INCLUDE CHILDREN. This will cause ColorPickers to be moved into the rig object.
+        move_into_collection(LightingPanelNames.Objects.LIGHTING_PANEL, char_name, include_children=False)
+        bpy.data.collections.remove(bpy.data.collections.get(LightingPanelNames.Collections.LIGHTING_PANEL), do_unlink=True)
+
     # If it exists, gets rid of the default collection.
     camera_coll = bpy.data.collections.get("Collection")
     if camera_coll:
         bpy.data.collections.remove(camera_coll,do_unlink=True)
 
-    # Let's 'exclude' that wgt collection: https://blenderartists.org/t/disable-exlude-from-view-layer-in-collection/1324744
-    def recurLayerCollection(layerColl):
-        found = None
-        if (layerColl.name == "wgt"):
-            return layerColl
-        for layer in layerColl.children:
-            found = recurLayerCollection(layer)
-            if found:
-                return found
     layer_collection = bpy.context.view_layer.layer_collection
-    layerColl = recurLayerCollection(layer_collection)
+    layerColl = searchForLayerCollection(layer_collection, "wgt")
     bpy.context.view_layer.active_layer_collection = layerColl
 
     layerColl.exclude = True
@@ -1094,7 +1106,12 @@ def rig_character(
     face_rig_obj = bpy.data.objects.get("facerig")
     if face_rig_obj:
         face_rig_obj.select_set(True)
-        
+
+    # Select lighting panel armature
+    lighting_panel_rig_obj = bpy.data.objects.get(LightingPanelNames.Objects.LIGHTING_PANEL)
+    if lighting_panel_rig_obj:
+        lighting_panel_rig_obj.select_set(True)
+
     # Select eye rig    
     eye_rig_obj = bpy.data.objects.get("eyerig")
     if eye_rig_obj:
@@ -1174,6 +1191,8 @@ def rig_character(
     bpy.ops.object.mode_set(mode='EDIT')
     armature.edit_bones['plate-border'].parent = armature.edit_bones['head']
     armature.edit_bones['plate-settings'].parent = armature.edit_bones['head']
+    if armature.edit_bones.get(LightingPanelNames.Bones.LIGHTING_PANEL):
+        armature.edit_bones[LightingPanelNames.Bones.LIGHTING_PANEL].parent = armature.edit_bones['head']
     
     armature.edit_bones['plate-border'].head = armature.edit_bones['neck'].head.copy()
     armature.edit_bones['plate-border'].tail = armature.edit_bones['neck'].tail.copy()
@@ -2576,6 +2595,8 @@ def rig_character(
         collections.new("Root")
         collections.new("Physics")
         collections.new("Cage")
+        if lighting_panel_rig_obj:
+            collections.new("Lighting")
         collections.new("Other")
         
         for bone in armature.bones:
@@ -2639,7 +2660,14 @@ def rig_character(
     rig_text = rig_file.as_string()
     complete_rig_text = rig_text
     # My disclaimer, out of respect for modifying Rigify core's script
-    rig_text_disclaimer = "\n# This RigUI script has been modified by Llama for use with Genshin Impact characters using a custom made rig. Any issues arising as a result of these modifications are not indicitive of Rigify's original functionalities. \n# Rigify's writers bare no responsibility for issues/errors made here. Additionally, these modifications have been made to improve the custom made rigs for Genshin Impact characters, meaning that\n# attempting to use this script elsewhere for characters/models/skeletons it was NOT intended to be used with, could yield improper or erroneous results - of which neither Rigify's development team nor I, am responsible for. \n\n# Otherwise, if you are seeing this disclaimer with a Hoyoverse character made with the proper addons, run this as needed. (Such as after appending to build the rig layers)\n# Do NOT however, attempt to use this rig in another version of blender than what it was made in. (3.6.X rigs will NOT work adequately in 4.X or beyond; and 4.X rigs will not work in previous versions before 4.0)\n"
+    rig_text_disclaimer = """
+# This RigUI script has been modified by Llama for use with Genshin Impact characters using a custom-made rig. Any issues arising as a result of these modifications are not indicative of Rigify's original functionalities.
+# Rigify's authors bear no responsibility for issues/errors resulting from these modifications. Additionally, these modifications have been made to improve the custom-made rigs for Genshin Impact characters.
+# Attempting to use this script with characters/models/skeletons it was not intended for could yield improper or erroneous results, for which neither Rigify's development team nor I are responsible.
+
+# If you are seeing this disclaimer on a Hoyoverse character made with the proper addons, run this as needed (e.g., after appending to build the rig layers).
+# Do NOT, however, attempt to use this rig in a version of Blender other than the one it was made in. (3.6.X rigs will NOT work adequately in 4.X or beyond, and 4.X rigs will not work in versions before 4.0).
+"""
     
     # MODIFICATIONS to the text file are made here:
     # Get the ID of this char's rig ui script.
@@ -2654,15 +2682,21 @@ def rig_character(
         else:
             return string3
     
+    # String object of the actual layers
     def layers_to_generate(vers):
-        str = "\n            row=col.row()\n            "+make_layer_str("Tweaks", 2, vers)+"\n            row=col.row()\n            "+make_layer_str("Pivots & Pins", 19, vers)+"\n            row = col.row()\n            "+make_layer_str("Offsets", 26, vers)+"\n            row = col.row()\n            "+make_layer_str("Props", 21, vers)+"\n            row = col.row()\n            row.separator()\n            row = col.row()\n            row.separator()\n            row = col.row()\n            "+make_layer_str("Face", 0, vers)+"\n            row = col.row()\n            "+make_layer_str("Torso (IK)", 3, vers)+"\n            row = col.row()\n            "+make_layer_str("Torso (FK)",4,vers)+"\n            row = col.row()\n            "+make_layer_str("Fingers", 5, vers)+"\n            row = col.row()\n            "+make_layer_str("Fingers (Detail)", 6, vers)+"\n            row = col.row()\n            "+make_layer_str("Arm.L (IK)", 7, vers)+"\n            "+make_layer_str("Arm.R (IK)", 10, vers)+"\n            row = col.row()\n            "+make_layer_str("Arm.L (FK)", 8, vers)+"\n            "+make_layer_str("Arm.R (FK)", 11, vers)+"\n            row = col.row()\n            "+make_layer_str("Leg.L (IK)", 13, vers)+"\n            "+make_layer_str("Leg.R (IK)", 16, vers)+"\n            row = col.row()\n            "+make_layer_str("Leg.L (FK)", 14, vers)+"\n            "+make_layer_str("Leg.R (FK)", 17, vers)+"\n            row = col.row()\n            row.separator()\n            row = col.row()\n            row.separator()\n            row = col.row()\n            "+make_layer_str("Root", 28, vers)+"\n            row = col.row()\n            "+make_layer_str("Physics", 20, vers)+"\n            row = col.row()\n            "+make_layer_str("Cage", 24, vers)+"\n            "+make_layer_str("Other", 25, vers)
+        str = "\n            row=col.row()\n            "+make_layer_str("Tweaks", 2, vers)+"\n            row=col.row()\n            "+make_layer_str("Pivots & Pins", 19, vers)+"\n            row = col.row()\n            "+make_layer_str("Offsets", 26, vers)+"\n            row = col.row()\n            "+make_layer_str("Props", 21, vers)+"\n            row = col.row()\n            row.separator()\n            row = col.row()\n            row.separator()\n            row = col.row()\n            "+make_layer_str("Face", 0, vers)+"\n            row = col.row()\n            "+make_layer_str("Torso (IK)", 3, vers)+"\n            row = col.row()\n            "+make_layer_str("Torso (FK)",4,vers)+"\n            row = col.row()\n            "+make_layer_str("Fingers", 5, vers)+"\n            row = col.row()\n            "+make_layer_str("Fingers (Detail)", 6, vers)+"\n            row = col.row()\n            "+make_layer_str("Arm.L (IK)", 7, vers)+"\n            "+make_layer_str("Arm.R (IK)", 10, vers)+"\n            row = col.row()\n            "+make_layer_str("Arm.L (FK)", 8, vers)+"\n            "+make_layer_str("Arm.R (FK)", 11, vers)+"\n            row = col.row()\n            "+make_layer_str("Leg.L (IK)", 13, vers)+"\n            "+make_layer_str("Leg.R (IK)", 16, vers)+"\n            row = col.row()\n            "+make_layer_str("Leg.L (FK)", 14, vers)+"\n            "+make_layer_str("Leg.R (FK)", 17, vers)+"\n            row = col.row()\n            row.separator()\n            row = col.row()\n            row.separator()\n            row = col.row()\n            "+make_layer_str("Root", 28, vers)
+        
+        if lighting_panel_rig_obj:
+            str+="\n            row = col.row()\n            "+make_layer_str("Lighting", 1, vers)
+        
+        str+="\n            row = col.row()\n            "+make_layer_str("Physics", 20, vers)+"\n            row = col.row()\n            "+make_layer_str("Cage", 24, vers)+"\n            "+make_layer_str("Other", 25, vers)
                 
         return str
         
     # Function to add layer to rigUI. This should add it to both 3.6 and 4.0 versions of the UI.
     def generate_rig_layers():
         # Add the physics button to the UI # text=v_str+" rig for " + char_name
-        rig_add_layer_code = "\n        layout = self.layout\n        col = layout.column()\n        row = col.row()\n        v_str = \""+bpy.app.version_string+"\"\n        if not v_str[0] == \"4\" and bpy.app.version_string[0] == \"3\":\n            row.label(text=v_str+\" rig for "+char_name.split("Costume")[0]+"\")\n            "+layers_to_generate(3)+"\n        elif v_str[0] == \"4\" and bpy.app.version_string[0] == \"4\":\n            row.label(text=v_str+\" rig for "+char_name.split("Costume")[0]+"\")\n            # If you have duplicate armatures of the same character (if you see .001 or similar) in one scene,\n            # Please change the name below to what it is in the Outliner so that you can rig all your characters :)\n            # (It's the green person symbol in your rig)\n            collection = bpy.data.armatures[\""+original_name+"\"].collections\n            "+layers_to_generate(4)+"\n        else:\n            row.label(text=\"ERROR: Version mismatch!\")\n            row = col.row()\n            row.label(text=\"Your rig was made in a version of Blender/Goo Engine that is not compatible!\")\n            row = col.row()\n            row.label(text=\"Please remake your rig for this version!\")"
+        rig_add_layer_code = "\n        layout = self.layout\n        col = layout.column()\n        row = col.row()\n        v_str = \""+bpy.app.version_string+"\"\n        row.label(text=v_str+\" rig for "+char_name.split("Costume")[0]+"\")\n        if not v_str[0] == \"4\" and bpy.app.version_string[0] == \"3\":\n            "+layers_to_generate(3)+"\n        elif v_str[0] == \"4\" and bpy.app.version_string[0] == \"4\":\n            # If you have duplicate armatures of the same character (if you see .001 or similar) in one scene,\n            # Please change the name below to what it is in the Outliner so that you can rig all your characters :)\n            # (It's the green person symbol in your rig)\n            collection = bpy.data.armatures[\""+original_name+"\"].collections\n            "+layers_to_generate(4)+"\n        else:\n            row.label(text=\"ERROR: Version mismatch!\")\n            row = col.row()\n            row.label(text=\"Your rig was made in a version of Blender/Goo Engine that is not compatible!\")\n            row = col.row()\n            row.label(text=\"Please remake your rig for this version!\")"
         cut_rig_layer = rig_text.split("class RigLayers(bpy.types.Panel):")
         separate_draw_func = cut_rig_layer[1].split("def draw(self, context):")
         separate_draw_end = separate_draw_func[1].split("def register():")
@@ -2746,6 +2780,33 @@ def rig_character(
     
     # DONE MODIFYING ui.py FILE --------------------------------------------
     
+    
+    # if using lighting panel, tie the visiblity of the RGB circle meshes to the visibility of the lighting layer.
+    if lighting_panel_rig_obj:
+        def drive_visibility_with_prop(obj, path):
+            driver_obj = bpy.context.scene.objects[obj]
+            driver = driver_obj.driver_add("hide_viewport").driver
+            
+            driver.type = 'SCRIPTED'
+            driver.expression = 'not is_visible'
+            
+            var = driver.variables.new()
+            var.name = "is_visible"
+            var.type = "SINGLE_PROP"
+            var.targets[0].id_type = "ARMATURE"
+            var.targets[0].id = armature
+            var.targets[0].data_path = path
+            
+        drive_visibility_with_prop("ColorWheel-Ambient","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-Fresnel","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-Lit","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-RimLit","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-RimShadow","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-Shadow","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-SoftLit","collections[\"Lighting\"].is_visible")
+        drive_visibility_with_prop("ColorWheel-SoftShadow","collections[\"Lighting\"].is_visible")
+  
+    
     # Post modification, Adjustment of bone layers/collections.
     if not is_version_4:
         for x in range(29):
@@ -2754,6 +2815,7 @@ def rig_character(
                 
         # Disable/Enable Rig UI layers we care about
         bpy.context.object.data.layers[0] = True
+        bpy.context.object.data.layers[1] = True if lighting_panel_rig_obj else False  # Lighting
         bpy.context.object.data.layers[3] = True
         bpy.context.object.data.layers[4] = False
         bpy.context.object.data.layers[5] = True
@@ -2949,7 +3011,28 @@ def rig_character(
     bone_to_layer("f_pinky.03.R", 6, "Fingers (Detail)")  
     bone_to_layer("f_pinky.01.L.001", 6, "Fingers (Detail)")  
     bone_to_layer("f_pinky.01.R.001", 6, "Fingers (Detail)") 
-    
+
+    if lighting_panel_rig_obj:
+        bone_to_layer("Lighting Panel", 1, "Lighting")
+        bone_to_layer("Fresnel", 1, "Lighting")
+        bone_to_layer("Ambient", 1, "Lighting")
+        bone_to_layer("SoftLit", 1, "Lighting")
+        bone_to_layer("Lit", 1, "Lighting")  # Sharp Lit
+        bone_to_layer("SoftShadow", 1, "Lighting")
+        bone_to_layer("Shadow", 1, "Lighting")  # Sharp Shadow
+        bone_to_layer("RimShadow", 1, "Lighting")
+        bone_to_layer("Rim Lit", 1, "Lighting")
+        bone_to_layer("RimX", 1, "Lighting")
+        bone_to_layer("RimY", 1, "Lighting")
+        bone_to_layer("RimLitPin", 1, "Lighting")
+        bone_to_layer("ShadowPin", 1, "Lighting")
+        bone_to_layer("LitPin", 1, "Lighting")
+        bone_to_layer("AmbientPin", 1, "Lighting")
+        bone_to_layer("RimShadowPin", 1, "Lighting")
+        bone_to_layer("SoftShadowPin", 1, "Lighting")
+        bone_to_layer("SoftLitPin", 1, "Lighting")
+        bone_to_layer("FresnelPin", 1, "Lighting")
+
     # Pass in a list, all of those bones will be moved accordingly.
     def fast_bone_move(bone_list, layer, collection):
         for bone in bone_list:
@@ -3109,3 +3192,39 @@ def setup_finger_scale_controls_on_x_axis_to_curl_just_the_fingertips(rigified_r
     for side in [".L", ".R"]:
         for bone in fingerlist:
             rigified_rig.pose.bones[bone + side].lock_scale[0] = False
+
+
+# Let's 'exclude' that wgt collection: https://blenderartists.org/t/disable-exlude-from-view-layer-in-collection/1324744
+def searchForLayerCollection(layerColl, coll_name):
+    found = None
+    if (layerColl.name == coll_name):
+        return layerColl
+    for layer in layerColl.children:
+        found = searchForLayerCollection(layer, coll_name)
+        if found:
+            return found
+
+
+def searchForParentLayerCollection(layerColl, coll_name):
+    found = None
+    for layer in layerColl.children:
+        if (layer.name == coll_name):
+            return layerColl
+        found = searchForParentLayerCollection(layer, coll_name)
+        if found:
+            return found
+
+
+def disable_collection(collection_name):
+    view_layer_collection = bpy.context.view_layer.layer_collection
+
+    layer_collection_to_disable = searchForLayerCollection(view_layer_collection, collection_name)
+    if layer_collection_to_disable:
+        layer_collection_to_disable.exclude = True
+        return True
+    return False
+
+
+def move_collection_into_collection(source, destination, collection):  
+    destination.children.link(collection)
+    source.children.unlink(collection)
